+++ packages/stream_chat_flutter/lib/src/message_widget/components/stream_message_content.dart
@@ -141,6 +141,7 @@ class _StreamMessageContentState extends State<StreamMessageContent> {
     final attachmentsWidth = renderBox.size.width;
 
     if (attachmentsWidth == 0) return;
+    if (widthLimit == attachmentsWidth) return;
     setState(() => widthLimit = attachmentsWidth);
   }